+++ packages/attach/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.7](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-attach@4.0.6...@alfalab/core-components-attach@4.0.7) (2022-02-15)
+
+**Note:** Version bump only for package @alfalab/core-components-attach
+
+
+
+
+
+## [4.0.6](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-attach@4.0.5...@alfalab/core-components-attach@4.0.6) (2022-02-09)
+
+**Note:** Version bump only for package @alfalab/core-components-attach
+
+
+
+
+
 ## [4.0.5](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-attach@4.0.4...@alfalab/core-components-attach@4.0.5) (2022-02-03)
 
 **Note:** Version bump only for package @alfalab/core-components-attach