+++ packages/custom-button/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-custom-button@1.2.0...@alfalab/core-components-custom-button@1.2.1) (2022-02-15)
+
+**Note:** Version bump only for package @alfalab/core-components-custom-button
+
+
+
+
+
+# [1.2.0](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-custom-button@1.1.3...@alfalab/core-components-custom-button@1.2.0) (2022-02-09)
+
+
+### Features
+
+* **themes:** introducing intranet theme ([#983](https://github.com/alfa-laboratory/core-components/issues/983)) ([85eb9cf](https://github.com/alfa-laboratory/core-components/commit/85eb9cfffeef31b886c5123d6333e177c261ac62))
+
+
+
+
+
 ## [1.1.3](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-custom-button@1.1.2...@alfalab/core-components-custom-button@1.1.3) (2022-02-03)