+++ packages/gallery/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.9](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-gallery@3.0.8...@alfalab/core-components-gallery@3.0.9) (2022-02-15)
+
+
+### Bug Fixes
+
+* **gallery:** update header button's tooltip position ([#979](https://github.com/alfa-laboratory/core-components/issues/979)) ([8e2acfc](https://github.com/alfa-laboratory/core-components/commit/8e2acfcb772cf4d8051c185a3f1caaaf507b5bc9))
+
+
+
+
+
+## [3.0.8](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-gallery@3.0.7...@alfalab/core-components-gallery@3.0.8) (2022-02-09)
+
+**Note:** Version bump only for package @alfalab/core-components-gallery
+
+
+
+
+
 ## [3.0.7](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-gallery@3.0.6...@alfalab/core-components-gallery@3.0.7) (2022-02-03)
 
 **Note:** Version bump only for package @alfalab/core-components-gallery