+++ packages/masked-input/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.13](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-masked-input@4.2.12...@alfalab/core-components-masked-input@4.2.13) (2022-02-15)
+
+**Note:** Version bump only for package @alfalab/core-components-masked-input
+
+
+
+
+
+## [4.2.12](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-masked-input@4.2.11...@alfalab/core-components-masked-input@4.2.12) (2022-02-09)
+
+**Note:** Version bump only for package @alfalab/core-components-masked-input
+
+
+
+
+
 ## [4.2.11](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-masked-input@4.2.10...@alfalab/core-components-masked-input@4.2.11) (2022-02-03)
 
 **Note:** Version bump only for package @alfalab/core-components-masked-input