+++ packages/pagination/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.4](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-pagination@1.1.3...@alfalab/core-components-pagination@1.1.4) (2022-02-15)
+
+**Note:** Version bump only for package @alfalab/core-components-pagination
+
+
+
+
+
+## [1.1.3](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-pagination@1.1.2...@alfalab/core-components-pagination@1.1.3) (2022-02-09)
+
+**Note:** Version bump only for package @alfalab/core-components-pagination
+
+
+
+
+
 ## [1.1.2](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-pagination@1.1.1...@alfalab/core-components-pagination@1.1.2) (2022-02-03)
 
 **Note:** Version bump only for package @alfalab/core-components-pagination