+++ packages/themes/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.4.0](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-themes@7.3.0...@alfalab/core-components-themes@7.4.0) (2022-02-16)
+
+
+### Features
+
+* **code-input:** add component ([#932](https://github.com/alfa-laboratory/core-components/issues/932)) ([dc40cb5](https://github.com/alfa-laboratory/core-components/commit/dc40cb5b28322b4a2dc5735b354a7d45cf34adb9))
+
+
+
+
+
+# [7.3.0](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-themes@7.2.1...@alfalab/core-components-themes@7.3.0) (2022-02-09)
+
+
+### Features
+
+* **themes:** introducing intranet theme ([#983](https://github.com/alfa-laboratory/core-components/issues/983)) ([85eb9cf](https://github.com/alfa-laboratory/core-components/commit/85eb9cfffeef31b886c5123d6333e177c261ac62))
+
+
+
+
+
 ## [7.2.1](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-themes@7.2.0...@alfalab/core-components-themes@7.2.1) (2022-02-02)