+++ packages/toast-plate/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.17](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-toast-plate@4.0.16...@alfalab/core-components-toast-plate@4.0.17) (2022-02-15)
+
+**Note:** Version bump only for package @alfalab/core-components-toast-plate
+
+
+
+
+
+## [4.0.16](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-toast-plate@4.0.15...@alfalab/core-components-toast-plate@4.0.16) (2022-02-09)
+
+**Note:** Version bump only for package @alfalab/core-components-toast-plate
+
+
+
+
+
 ## [4.0.15](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-toast-plate@4.0.14...@alfalab/core-components-toast-plate@4.0.15) (2022-02-03)
 
 **Note:** Version bump only for package @alfalab/core-components-toast-plate