+++ packages/with-suffix/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.13](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-with-suffix@3.1.12...@alfalab/core-components-with-suffix@3.1.13) (2022-02-15)
+
+**Note:** Version bump only for package @alfalab/core-components-with-suffix
+
+
+
+
+
+## [3.1.12](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-with-suffix@3.1.11...@alfalab/core-components-with-suffix@3.1.12) (2022-02-09)
+
+**Note:** Version bump only for package @alfalab/core-components-with-suffix
+
+
+
+
+
 ## [3.1.11](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-with-suffix@3.1.10...@alfalab/core-components-with-suffix@3.1.11) (2022-02-03)
 
 **Note:** Version bump only for package @alfalab/core-components-with-suffix